Resolve import cycle by moving VCommand to a new module
