+++ varats-core/varats/project/project_util.py
@@ -7,16 +7,12 @@ from pathlib import Path
 
 import benchbuild as bb
 import pygit2
-from benchbuild.command import Command
 from benchbuild.source import Git
 from benchbuild.utils.cmd import git
 from plumbum import local
 from plumbum.commands.base import BoundCommand
 
 from varats.utils.settings import bb_cfg
-
-if tp.TYPE_CHECKING:
-    from varats.provider.patch.patch_provider import Patch, PatchSet
 
 LOG = logging.getLogger(__name__)
 
@@ -386,87 +382,3 @@ def copy_renamed_git_to_dest(src_dir: Path, dest_dir: Path) -> None:
         for name in dirs:
             if name == ".gitted":
                 os.rename(os.path.join(root, name), os.path.join(root, ".git"))
-
-
-class VCommand(Command):  # type: ignore [misc]
-    """
-    Wrapper around benchbuild's Command class.
-
-    Attributes:
-    requires_any_args: any of these command line args must be available for
-                       successful execution.
-    requires_all_args: all of these command line args must be available for
-                       successful execution.
-    requires_any_patch: any of these patch feature-tags must be available for
-                       successful execution.
-    requires_all_patch: all of these patch feature-tags must be available for
-                       successful execution.
-    """
-
-    _requires: tp.Set[str]
-
-    def __init__(
-        self,
-        *args: tp.Any,
-        requires_any_args: tp.Optional[tp.Set[str]] = None,
-        requires_all_args: tp.Optional[tp.Set[str]] = None,
-        requires_any_patch: tp.Optional[tp.Set[str]] = None,
-        requires_all_patch: tp.Optional[tp.Set[str]] = None,
-        **kwargs: tp.Union[str, tp.List[str]],
-    ) -> None:
-
-        super().__init__(*args, **kwargs)
-        self._requires_any_args = requires_any_args or set()
-        self._requires_all_args = requires_all_args or set()
-        self._requires_any_patch = requires_any_patch or set()
-        self._requires_all_patch = requires_all_patch or set()
-
-    @property
-    def requires_any_args(self) -> tp.Set[str]:
-        return self._requires_any_args
-
-    @property
-    def requires_all_args(self) -> tp.Set[str]:
-        return self._requires_all_args
-
-    @property
-    def requires_any_patch(self) -> tp.Set[str]:
-        return self._requires_any_patch
-
-    @property
-    def requires_all_patch(self) -> tp.Set[str]:
-        return self._requires_all_patch
-
-    def can_be_executed_by(
-        self, extra_args: tp.Set[str], applied_patches: 'PatchSet'
-    ) -> bool:
-        """
-        Checks whether this command can be executed with the give configuration.
-
-        Args:
-            extra_args: additional command line arguments that will be passed to
-                        the command
-            applied_patches: patches that were applied to create the executable
-
-        Returns:
-            whether this command can be executed
-        """
-        all_args = set(self._args).union(extra_args)
-        all_patch_tags: tp.Set[str] = set()
-        for patch in applied_patches:
-            if patch.feature_tags:
-                all_patch_tags.update(patch.feature_tags)
-
-        return bool((
-            not self.requires_any_args or
-            all_args.intersection(self.requires_any_args)
-        ) and (
-            not self.requires_all_args or
-            self.requires_all_args.issubset(all_args)
-        ) and (
-            not self.requires_any_patch or
-            all_patch_tags.intersection(self.requires_any_patch)
-        ) and (
-            not self.requires_all_patch or
-            self.requires_all_patch.issubset(all_patch_tags)
-        ))
